Fixed not getting attributes through TvSeries
Also corrected references to `Person.involvement` with `Person.involvements`, and corrected references to `Title.genre`.

Also replaced some conditions for edges to connect nodes with simply `imdb.titles`, as this connects the main node to every other title in the diagram, which is the case for Shared Crew Diagram and Persons' Titles Diagram.